+++ packages/core/src/helpers/is-root-text-node.ts
@@ -3,7 +3,7 @@ import { MitosisNode } from '../types/mitosis-node';
 
 export function isRootTextNode(json: MitosisComponent | MitosisNode) {
   const firstChild = json.children[0];
-  return Boolean(firstChild && isTextNode(firstChild));
+  return Boolean(json.children.length === 1 && firstChild && isTextNode(firstChild));
 }
 
 export function isTextNode(node: MitosisNode) {