+++ packages/core/src/parsers/builder/builder.ts
@@ -1,10 +1,11 @@
+import { hashCodeAsString } from '@/symbols/symbol-processor';
+import { MitosisComponent, MitosisState } from '@/types/mitosis-component';
 import * as babel from '@babel/core';
 import generate from '@babel/generator';
 import { BuilderContent, BuilderElement } from '@builder.io/sdk';
 import json5 from 'json5';
 import { mapKeys, merge, omit, omitBy, sortBy, upperFirst } from 'lodash';
 import traverse from 'traverse';
-import { MitosisComponent, MitosisState, hashCodeAsString } from '../..';
 import { Size, sizeNames, sizes } from '../../constants/media-sizes';
 import { createSingleBinding } from '../../helpers/bindings';
 import { capitalize } from '../../helpers/capitalize';
@@ -128,6 +129,18 @@ const getStyleStringFromBlock = (block: BuilderElement, options: BuilderToMitosi
   }
 
   return styleString;
+};
+
+const hasComponent = (block: BuilderElement) => {
+  return Boolean(block.component?.name);
+};
+
+const hasProperties = (block: BuilderElement) => {
+  return Boolean(block.properties && Object.keys(block.properties).length);
+};
+
+const hasBindings = (block: BuilderElement) => {
+  return Boolean(block.bindings && Object.keys(block.bindings).length);
 };
 
 const hasStyles = (block: BuilderElement) => {
@@ -432,7 +445,13 @@ const componentMappers: {
     };
     const finalTagname = block.tagName || (assumeLink ? 'a' : 'div');
 
-    if ((block.tagName && block.tagName !== 'div') || hasStyles(block)) {
+    if (
+      (block.tagName && block.tagName !== 'div') ||
+      hasStyles(block) ||
+      hasComponent(block) ||
+      hasBindings(block) ||
+      hasProperties(block)
+    ) {
       return createMitosisNode({
         name: finalTagname,
         bindings,
@@ -465,6 +484,7 @@ type BuilderToMitosisOptions = {
   context?: { [key: string]: any };
   includeBuilderExtras?: boolean;
   preserveTextBlocks?: boolean;
+  includeSpecialBindings?: boolean;
 };
 
 export const builderElementToMitosisNode = (
@@ -472,6 +492,8 @@ export const builderElementToMitosisNode = (
   options: BuilderToMitosisOptions,
   _internalOptions: InternalOptions = {},
 ): MitosisNode => {
+  const { includeSpecialBindings = true } = options;
+
   if (block.component?.name === 'Core:Fragment') {
     block.component.name = 'Fragment';
   }
@@ -558,7 +580,9 @@ export const builderElementToMitosisNode = (
     return mapper(block, options);
   }
 
-  const bindings: any = {};
+  const bindings: MitosisNode['bindings'] = {};
+  const children: MitosisNode[] = [];
+  const slots: MitosisNode['slots'] = {};
 
   if (blockBindings) {
     for (const key in blockBindings) {
@@ -567,9 +591,9 @@ export const builderElementToMitosisNode = (
       }
       const useKey = key.replace(/^(component\.)?options\./, '');
       if (!useKey.includes('.')) {
-        bindings[useKey] = {
+        bindings[useKey] = createSingleBinding({
           code: (blockBindings[key] as any).code || blockBindings[key],
-        };
+        });
       } else if (useKey.includes('style') && useKey.includes('.')) {
         const styleProperty = useKey.split('.')[1];
         // TODO: add me in
@@ -599,10 +623,34 @@ export const builderElementToMitosisNode = (
   if (block.component?.options) {
     for (const key in block.component.options) {
       const value = block.component.options[key];
-      if (typeof value === 'string') {
+      const valueIsArrayOfBuilderElements = Array.isArray(value) && value.every(isBuilderElement);
+
+      const transformBldrElementToMitosisNode = (item: BuilderElement) => {
+        const node = builderElementToMitosisNode(item, {
+          ...options,
+          includeSpecialBindings: false,
+        });
+
+        return node;
+      };
+
+      if (isBuilderElement(value)) {
+        slots[key] = [transformBldrElementToMitosisNode(value)];
+      } else if (typeof value === 'string') {
         properties[key] = value;
+      } else if (valueIsArrayOfBuilderElements) {
+        const childrenElements = value
+          .filter((item) => {
+            if (item.properties?.src?.includes('/api/v1/pixel')) {
+              return false;
+            }
+            return true;
+          })
+          .map(transformBldrElementToMitosisNode);
+
+        slots[key] = childrenElements;
       } else {
-        bindings[key] = { code: json5.stringify(value) };
+        bindings[key] = createSingleBinding({ code: json5.stringify(value) });
       }
     }
   }
@@ -618,7 +666,7 @@ export const builderElementToMitosisNode = (
     if (binding.startsWith('component.options') || binding.startsWith('options')) {
       const value = blockBindings[binding];
       const useKey = binding.replace(/^(component\.options\.|options\.)/, '');
-      bindings[useKey] = { code: value };
+      bindings[useKey] = createSingleBinding({ code: value });
     }
   }
 
@@ -628,7 +676,7 @@ export const builderElementToMitosisNode = (
       block.tagName ||
       ((block as any).linkUrl ? 'a' : 'div'),
     properties: {
-      ...(block.component && { $tagName: block.tagName }),
+      ...(block.component && includeSpecialBindings && { $tagName: block.tagName }),
       ...(block.class && { class: block.class }),
       ...properties,
     },
@@ -636,12 +684,15 @@ export const builderElementToMitosisNode = (
       ...bindings,
       ...actionBindings,
       ...(styleString && {
-        style: { code: styleString },
+        style: createSingleBinding({ code: styleString }),
       }),
       ...(css &&
         Object.keys(css).length && {
-          css: { code: JSON.stringify(css) },
+          css: createSingleBinding({ code: JSON.stringify(css) }),
         }),
+    },
+    slots: {
+      ...slots,
     },
   });
 
@@ -677,7 +728,9 @@ export const builderElementToMitosisNode = (
     }
   }
 
-  node.children = (block.children || []).map((item) => builderElementToMitosisNode(item, options));
+  node.children = children.concat(
+    (block.children || []).map((item) => builderElementToMitosisNode(item, options)),
+  );
 
   return node;
 };